Split _type_results.py: extract execution-scoped types into _type_results_execution.py
Extract SessionTelemetry, ProviderOutcome, RecipeIdentity, CIRunScope into a new
peer module _type_results_execution.py. The re-export hub (types/__init__.py)
star-imports the new module so all downstream autoskillit.core import paths
continue to work. Test cascade for the new module targets 4 directories
(core, execution, server, pipeline) instead of the original 13.
